chore: make pipeline builder more object, almost done[part-4]
