+++ src/query/service/src/pipelines/builders/builder_window.rs
@@ -1,0 +1,137 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use common_exception::Result;
+use common_expression::with_number_mapped_type;
+use common_expression::SortColumnDescription;
+use common_pipeline_core::processors::processor::ProcessorPtr;
+use common_pipeline_core::processors::Processor;
+use common_sql::executor::Window;
+
+use crate::pipelines::processors::transforms::FrameBound;
+use crate::pipelines::processors::transforms::WindowFunctionInfo;
+use crate::pipelines::processors::TransformWindow;
+use crate::pipelines::PipelineBuilder;
+
+impl PipelineBuilder {
+    fn build_window(&mut self, window: &Window) -> Result<()> {
+        self.build_pipeline(&window.input)?;
+
+        let input_schema = window.input.output_schema()?;
+
+        let partition_by = window
+            .partition_by
+            .iter()
+            .map(|p| {
+                let offset = input_schema.index_of(&p.to_string())?;
+                Ok(offset)
+            })
+            .collect::<Result<Vec<_>>>()?;
+
+        let order_by = window
+            .order_by
+            .iter()
+            .map(|o| {
+                let offset = input_schema.index_of(&o.order_by.to_string())?;
+                Ok(SortColumnDescription {
+                    offset,
+                    asc: o.asc,
+                    nulls_first: o.nulls_first,
+                    is_nullable: input_schema.field(offset).is_nullable(), // Used for check null frame.
+                })
+            })
+            .collect::<Result<Vec<_>>>()?;
+
+        let old_output_len = self.main_pipeline.output_len();
+        if !partition_by.is_empty() || !order_by.is_empty() {
+            let mut sort_desc = Vec::with_capacity(partition_by.len() + order_by.len());
+
+            for offset in &partition_by {
+                sort_desc.push(SortColumnDescription {
+                    offset: *offset,
+                    asc: true,
+                    nulls_first: true,
+                    is_nullable: input_schema.field(*offset).is_nullable(),  // This information is not needed here.
+                })
+            }
+
+            sort_desc.extend(order_by.clone());
+
+            self.build_sort_pipeline(input_schema.clone(), sort_desc, window.plan_id, None, false)?;
+        }
+        // `TransformWindow` is a pipeline breaker.
+        self.main_pipeline.try_resize(1)?;
+        let func = WindowFunctionInfo::try_create(&window.func, &input_schema)?;
+        // Window
+        self.main_pipeline.add_transform(|input, output| {
+            // The transform can only be created here, because it cannot be cloned.
+
+            let transform = if window.window_frame.units.is_rows() {
+                let start_bound = FrameBound::try_from(&window.window_frame.start_bound)?;
+                let end_bound = FrameBound::try_from(&window.window_frame.end_bound)?;
+                Box::new(TransformWindow::<u64>::try_create_rows(
+                    input,
+                    output,
+                    func.clone(),
+                    partition_by.clone(),
+                    order_by.clone(),
+                    (start_bound, end_bound),
+                )?) as Box<dyn Processor>
+            } else {
+                if order_by.len() == 1 {
+                    // If the length of order_by is 1, there may be a RANGE frame.
+                    let data_type = input_schema
+                        .field(order_by[0].offset)
+                        .data_type()
+                        .remove_nullable();
+                    with_number_mapped_type!(|NUM_TYPE| match data_type {
+                        DataType::Number(NumberDataType::NUM_TYPE) => {
+                            let start_bound =
+                                FrameBound::try_from(&window.window_frame.start_bound)?;
+                            let end_bound = FrameBound::try_from(&window.window_frame.end_bound)?;
+                            return Ok(ProcessorPtr::create(Box::new(
+                                TransformWindow::<NUM_TYPE>::try_create_range(
+                                    input,
+                                    output,
+                                    func.clone(),
+                                    partition_by.clone(),
+                                    order_by.clone(),
+                                    (start_bound, end_bound),
+                                )?,
+                            )
+                                as Box<dyn Processor>));
+                        }
+                        _ => {}
+                    })
+                }
+
+                // There is no offset in the RANGE frame. (just CURRENT ROW or UNBOUNDED)
+                // So we can use any number type to create the transform.
+                let start_bound = FrameBound::try_from(&window.window_frame.start_bound)?;
+                let end_bound = FrameBound::try_from(&window.window_frame.end_bound)?;
+                Box::new(TransformWindow::<u8>::try_create_range(
+                    input,
+                    output,
+                    func.clone(),
+                    partition_by.clone(),
+                    order_by.clone(),
+                    (start_bound, end_bound),
+                )?) as Box<dyn Processor>
+            };
+            Ok(ProcessorPtr::create(transform))
+        })?;
+
+        self.main_pipeline.try_resize(old_output_len)
+    }
+}
